docker-compose/faithful/requirements.txt
absl-py==0.6.0
astor==0.7.1
attrs==18.2.0
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.0
colorlog==3.1.4
cycler==0.10.0
cymem==2.0.2
Cython==0.29
cytoolz==0.9.0.1
dill==0.2.8.2
easydev==0.9.37
enum34==1.1.6
env==0.1.0
# Flask==1.0.2
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
hypothesis==3.80.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==2.2.3
mock==2.0.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
murmurhash==1.0.1
nltk==3.3
numpy==1.15.3
opencv-python==3.4.3.18
pathlib==1.0.1
pbr==5.1.0
pexpect==4.6.0
Pillow==5.3.0
# pkg-resources==0.0.0
# Could not find a version that satisfies the requirement pkg-resources==0.0.0 (from -r /tmp/requirements.txt (line 47)) (from versions: )
# No matching distribution found for pkg-resources==0.0.0 (from -r /tmp/requirements.txt (line 47))
# You are using pip version 9.0.1, however version 18.1 is available.
# You should consider upgrading via the 'pip install --upgrade pip' command.

plac==0.9.6
preshed==2.0.1
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.17.1

pyemd==0.5.1
PyGObject==3.30.1
pyparsing==2.2.2
python-dateutil==2.7.3
python-env-utils==0.4.1
pytz==2018.5
PyYAML==3.13
regex==2018.1.10
requests==2.20.0
rfc3987==1.3.8
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
sklearn==0.0
spacy==2.0.16
subprocess32==3.5.3
tensorboard==1.11.0
tensorboardX==1.4
tensorflow==1.11.0
termcolor==1.1.0
thinc==6.12.0
toolz==0.9.0
torch==0.4.1
torchvision==0.2.1
tqdm==4.27.0
typing==3.6.6
ujson==1.35
urllib3==1.24
utils==0.9.0
Werkzeug==0.14.1
wrapt==1.10.11

benepar==0.0.3
pycocotools==2.0.0